Add files via upload
Basically we were given username , password and initial balance in account which was already stored in a Google spreadsheet we had to read that data and through stored login credentials we had to login into our account. For that I used function of Appscript to do it.After login I have given the user choice to tell if it wants to see balance, credit the money or debit and accordingly whatever the user chooses the bot takes it further . If the user chooses debit he will be asked to enter the amount in multiples of 100 to debit , Once he input the no of notes of 100 the bot will show total amount debited and remaining balance in the account , it will also be printed on serial monitor and this data will also be sent to seperate google sheets.\\Similarly happens with credit command when we give the command it ask us for the no of notes of 100rs and after we enter it final balance after credission and amount credited in account will be shown on serial monitor. Now we were given that whatever transaction we do should be displayed on g.sheet and should be updated on it frequently so we made a seperate function transactionto carry which makes us allow our output to store in g.sheet. Now when the data is stored we will be able to see this on g.sheet.